requirements.txt
python-dotenv==1.0.1
uvicorn==0.32.0
fastapi==0.115.5
pandas==2.2.3
